프로필 이미지 업로드 과정 삭제 로직 수정

+++ DontBeServer/src/main/java/com/dontbe/www/DontBeServer/api/member/service/MemberCommandService.java
@@ -146,7 +146,7 @@ public class MemberCommandService {
                 String s3ImageUrl = s3Service.uploadImage(memberId.toString(), multipartFile);
                 existingMember.updateProfileUrl(s3ImageUrl);
 
-                if(!existedImage.equals(GHOST_IMAGE)||!existedImage.equals(GHOST_IMAGE_S3)) {
+                if(!existedImage.equals(GHOST_IMAGE)&&!existedImage.equals(GHOST_IMAGE_S3)) {
                     String existedKey = removeBaseUrl(existedImage, S3_URL);
                     s3Service.deleteImage(existedKey);
                 }
